Add files via upload
To automate typing into text areas with blocked copy-paste options. 
Enter all the text to be typed in a your_file.txt and run the program.
After 5 seconds the program starts typing so place your cursor in desired window.
Does not work in background.
For IDEs, closing brackets are provided automatically so delete the extra brackets at the end before compiling code etc.